+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Some tools (data generators, encoders/decoders and text case conversion) are now also available in the Editor menu or code intentions. Some of these actions are only available if a text is selected, or the current caret position is on a Java/Kotlin string or identifier.
 - Extend the ULID generator for monotonic ULIDs.
 
 ### Changed